Add interactive gti inspection tool
Adds a interactive gti inspection tool. Given a list of event files, the
tool will construct a light curve and plot it with the GTI intervals
as an overlay. Though a click-and-drag interface the user can then
interactive add or remove to the GTI table. Changed made are stored to
the event file internal GTI table.